Fix CI build: Move EspressoIdlingResource to androidTest
Espresso test dependencies are only available in androidTest, not debug.
Moving the IdlingResource implementation to uitests/androidTest resolves
compilation errors while maintaining test functionality.
